update comments; add a few tests

+++ content/PublishPackageToPowerShellGallery/PublishPackageToPowerShellGallery.ps1
@@ -3,9 +3,11 @@
 param(
     [parameter(Mandatory = $true)] [string] $apiKey,
     [parameter(Mandatory = $true)] [string] $path,
-    [parameter(Mandatory = $true)] [string] $setVersionNumberInManifest,
+    [parameter(Mandatory = $false)] [string] $setVersionNumberInManifest,
     [parameter(Mandatory = $false)] [string] $psd1FileName,
-    [parameter(Mandatory = $false)] [string] $version
+    [parameter(Mandatory = $false)] [string] $version,
+    [parameter(Mandatory = $false)] [switch] $whatifpublish,
+    [parameter(Mandatory = $false)] [switch] $whatifedit
 )
 Function Edit-ModuleVersionNumber {
     [cmdletbinding()]
@@ -13,8 +15,10 @@ Function Edit-ModuleVersionNumber {
         [ValidateNotNullOrEmpty()]
         $ModuleVersionNumber,
         [ValidateNotNullOrEmpty()]
-        $psd1File
+        $psd1File, 
+        [switch] $whatif
     )
+    $psd1File = Resolve-Path $psd1File
     $ModuleVersionNumber = $ModuleVersionNumber.Trim()
     if ((Test-Path $psd1File) -eq $false) {
         Write-Error "$psd1File does not exist!"
@@ -42,16 +46,19 @@ Function Edit-ModuleVersionNumber {
         }
     }
     Write-Host "VersionNumber in $psd1Name will be $ModuleVersionNumber."
-    try {
-        $LineToUpdate = Select-String -Pattern 'ModuleVersion(.*)' -Path $psd1File | Select-Object -First 1
-        (Get-Content $psd1File) -replace $LineToUpdate.line, "ModuleVersion = '$ModuleVersionNumber'" | Set-Content $psd1File
-        [string]$updatedModuleVersion = Get-Content $psd1File | Where-Object { $_ -match "ModuleVersion" } | Select-Object -First 1
-        $updatedModuleVersion = $updatedModuleVersion.Trim()
-        Write-Host "Updated to $updatedModuleVersion"
-    }
-    catch {
-        Write-Error "Something went wrong in updating ModuleNumber."
-        Throw $_.Exception
+    
+    if ($PSBoundParameters.ContainsKey('whatif') -eq $false) {
+        try {
+            $LineToUpdate = Select-String -Pattern 'ModuleVersion(.*)' -Path $psd1File | Select-Object -First 1
+            (Get-Content $psd1File) -replace $LineToUpdate.line, "ModuleVersion = '$ModuleVersionNumber'" | Set-Content $psd1File
+            [string]$updatedModuleVersion = Get-Content $psd1File | Where-Object { $_ -match "ModuleVersion" } | Select-Object -First 1
+            $updatedModuleVersion = $updatedModuleVersion.Trim()
+            Write-Host "Updated to $updatedModuleVersion"
+        }
+        catch {
+            Write-Error "Something went wrong in updating ModuleNumber."
+            Throw $_.Exception
+        }
     }
 }
 Function Publish-PackageToPowerShellGallery {
@@ -60,42 +67,60 @@ Function Publish-PackageToPowerShellGallery {
         [ValidateNotNullOrEmpty()]
         $apiKey,
         [ValidateNotNullOrEmpty()]
-        $path
+        $path, 
+        [switch] $whatif
     )
-    $path = Resolve-Path $path
-    Write-Host $path
-    $nugetPath = "c:\nuget"
-    if (!(Test-Path -Path $nugetPath)) {
-        Write-Verbose "Creating directory $nugetPath" -Verbose
-        New-Item -Path $nugetPath -ItemType Directory
-    }
-    Write-Verbose "Working Folder : $nugetPath"
-    $NugetExe = "$nugetPath\nuget.exe"
-    if (-not (Test-Path $NugetExe)) {
-        Write-Verbose "Cannot find nuget at $NugetExe" -Verbose
-        $NuGetInstallUri = "https://dist.nuget.org/win-x86-commandline/latest/nuget.exe"
-        $sourceNugetExe = $NuGetInstallUri
-        Write-Verbose "$sourceNugetExe -OutFile $NugetExe" -Verbose
-        Invoke-WebRequest $sourceNugetExe -OutFile $NugetExe
-        if (-not (Test-Path $NugetExe)) { 
-            Throw "Nuget download hasn't worked."
+    
+    if ($PSBoundParameters.ContainsKey('whatif') -eq $false) {
+        $path = Resolve-Path $path
+        $nugetPath = "c:\nuget"
+        if (!(Test-Path -Path $nugetPath)) {
+            Write-Verbose "Creating directory $nugetPath" -Verbose
+            New-Item -Path $nugetPath -ItemType Directory
         }
-        Else {Write-Verbose "Nuget Downloaded!" -Verbose}
+        Write-Verbose "Working Folder : $nugetPath"
+        $NugetExe = "$nugetPath\nuget.exe"
+        if (-not (Test-Path $NugetExe)) {
+            Write-Verbose "Cannot find nuget at $NugetExe" -Verbose
+            $NuGetInstallUri = "https://dist.nuget.org/win-x86-commandline/latest/nuget.exe"
+            $sourceNugetExe = $NuGetInstallUri
+            Write-Verbose "$sourceNugetExe -OutFile $NugetExe" -Verbose
+        
+            Invoke-WebRequest $sourceNugetExe -OutFile $NugetExe
+            if (-not (Test-Path $NugetExe)) { 
+                Throw "Nuget download hasn't worked."
+            }
+            Else {Write-Verbose "Nuget Downloaded!" -Verbose}
+        }
+        Write-Verbose "Add $nugetPath as %PATH%"
+        $pathenv = [System.Environment]::GetEnvironmentVariable("path")
+        $pathenv = $pathenv + ";" + $nugetPath
+        [System.Environment]::SetEnvironmentVariable("path", $pathenv)
+        Write-Verbose "Create NuGet package provider" -Verbose
+        Install-PackageProvider -Name NuGet -Scope CurrentUser -Force
+        Write-Verbose "Publishing module" -Verbose
+        Publish-Module -Path $path -NuGetApiKey $apiKey -Force
     }
-    Write-Verbose "Add $nugetPath as %PATH%"
-    $pathenv = [System.Environment]::GetEnvironmentVariable("path")
-    $pathenv = $pathenv + ";" + $nugetPath
-    [System.Environment]::SetEnvironmentVariable("path", $pathenv)
-    Write-Verbose "Create NuGet package provider" -Verbose
-    Install-PackageProvider -Name NuGet -Scope CurrentUser -Force
-    Write-Verbose "Publishing module" -Verbose
-    Publish-Module -Path $path -NuGetApiKey $apiKey -Force
+    else {
+        Publish-Module -Path $path -NuGetApiKey $apiKey -WhatIf
+    }
+}
+#tests
+if ($PSBoundParameters.ContainsKey('whatifpublish') -eq $true) {
+    Publish-PackageToPowerShellGallery -apiKey $apiKey -path $path -whatif   
+}
+if ($PSBoundParameters.ContainsKey('whatifedit') -eq $true) {
+    Edit-ModuleVersionNumber -ModuleVersionNumber $version -psd1File $psd1FileName -whatif
 }
 
-if ($setVersionNumberInManifest -eq $false) {
-    Publish-PackageToPowerShellGallery -apiKey $apiKey -path $path    
-}
-else {
-    Edit-ModuleVersionNumber -ModuleVersionNumber $version -psd1File $psd1FileName
-    Publish-PackageToPowerShellGallery -apiKey $apiKey -path $path
+#if both test switches are false, engage!
+if (($PSBoundParameters.ContainsKey('whatifedit') -eq $false) -and ($PSBoundParameters.ContainsKey('whatifpublish') -eq $false))
+{
+    if ($setVersionNumberInManifest -eq $false) {
+        Publish-PackageToPowerShellGallery -apiKey $apiKey -path $path    
+    }
+    else {
+        Edit-ModuleVersionNumber -ModuleVersionNumber $version -psd1File $psd1FileName
+        Publish-PackageToPowerShellGallery -apiKey $apiKey -path $path
+    }
 }
